Move field set (via handle) to top level class
