+++ src/main/java/io/quarkus/gizmo2/impl/FieldDeref.java
@@ -1,16 +1,12 @@
 package io.quarkus.gizmo2.impl;
 
-import static java.lang.constant.ConstantDescs.*;
-
 import java.lang.constant.ClassDesc;
-import java.lang.constant.MethodTypeDesc;
 import java.util.function.BiFunction;
 
 import io.github.dmlloyd.classfile.CodeBuilder;
 import io.quarkus.gizmo2.InstanceFieldVar;
 import io.quarkus.gizmo2.MemoryOrder;
 import io.quarkus.gizmo2.desc.FieldDesc;
-import io.quarkus.gizmo2.impl.constant.ConstImpl;
 
 public final class FieldDeref extends AssignableImpl implements InstanceFieldVar {
     private final Item item;
@@ -63,28 +59,7 @@ public final class FieldDeref extends AssignableImpl implements InstanceFieldVar
     Item emitSet(final BlockCreatorImpl block, final Item value, final MemoryOrder mode) {
         return switch (mode) {
             case AsDeclared -> new FieldSet(this, value);
-            default -> new Item() {
-                protected Node forEachDependency(Node node, final BiFunction<Item, Node, Node> op) {
-                    return ConstImpl.ofFieldVarHandle(desc)
-                            .process(io.quarkus.gizmo2.impl.FieldDeref.this.process(value.process(node.prev(), op), op), op);
-                }
-
-                public void writeCode(final CodeBuilder cb, final BlockCreatorImpl block) {
-                    cb.invokevirtual(CD_VarHandle, switch (mode) {
-                        case Plain -> "set";
-                        case Opaque -> "setOpaque";
-                        case Release -> "setRelease";
-                        case Volatile -> "setVolatile";
-                        default -> throw new IllegalStateException();
-                    }, MethodTypeDesc.of(
-                            desc().type(),
-                            Util.NO_DESCS));
-                }
-
-                public String itemName() {
-                    return io.quarkus.gizmo2.impl.FieldDeref.this.itemName() + ":set*";
-                }
-            };
+            default -> new FieldSetViaHandle(this, value, mode);
         };
     }
